+++ eras/alonzo/impl/src/Cardano/Ledger/Alonzo/Language.hs
@@ -1,8 +1,8 @@
 {-# LANGUAGE DeriveGeneric #-}
 
--- | This module exports the module Cardano.Ledger.Language. That module exports data
+-- | This module exports the module Cardano.Ledger.Plutus.Language. That module exports data
 --   structures and operations for talking about Non-native Script languages. It is expected that new languages (or new
 --   versions of old languages) will be added there, and just imported here.
-module Cardano.Ledger.Alonzo.Language (module Cardano.Ledger.Language) where
+module Cardano.Ledger.Alonzo.Language (module Cardano.Ledger.Plutus.Language) where
 
-import Cardano.Ledger.Language
+import Cardano.Ledger.Plutus.Language
