+++ eras/alonzo/test-suite/src/Test/Cardano/Ledger/Alonzo/Translation/Golden.hs
@@ -11,7 +11,7 @@ module Test.Cardano.Ledger.Alonzo.Translation.Golden (
 import Cardano.Ledger.Alonzo.TxInfo (ExtendedUTxO (..))
 import Cardano.Ledger.Binary.Encoding (serialize)
 import Cardano.Ledger.Core
-import Cardano.Ledger.Language (Language)
+import Cardano.Ledger.Plutus.Language (Language)
 import Control.Exception (throwIO)
 import qualified Data.ByteString.Lazy as BSL
 import Data.Functor.Identity (Identity)
