+++ libs/cardano-ledger-core/cardano-ledger-core.cabal
@@ -47,11 +47,12 @@ library
         Cardano.Ledger.Keys
         Cardano.Ledger.Keys.Bootstrap
         Cardano.Ledger.Keys.WitVKey
-        Cardano.Ledger.Language
+        --Cardano.Ledger.Language
         Cardano.Ledger.MemoBytes
         Cardano.Ledger.Orphans
         Cardano.Ledger.PoolDistr
         Cardano.Ledger.PoolParams
+        Cardano.Ledger.Plutus.Language
         Cardano.Ledger.Rewards
         Cardano.Ledger.Rules.ValidationMode
         Cardano.Ledger.SafeHash
